requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
faiss-cpu
yt-dlp
requests
python-dotenv
